* Finished initial work on splitting up special forms and scoping for various components/language dialects
