+++ lis.py
@@ -3,12 +3,16 @@ Lispy
 
 The main point of entry for lis.py, this file includes code to
 bootstrap the scripting language. It should not include implementation
-details for parsing or scoping
+details for parsing, scoping, special forms or types
 """
 import sys
 import traceback
+import argparse
 
-from lispy.norvig import read_from, tokenize, eval
+from lispy.dialects.norvig import read_from, tokenize, eval
+from lispy.dialects.norvig.scope import Scope, add_globals
+from lispy.dialects.norvig.special_forms import SPECIAL_FORMS as NORVIG_FORMS
+from lispy.dialects.pybuiltins.special_forms import SPECIAL_FORMS as PYBUILTIN_FORMS
 
 
 def read(s):
@@ -23,13 +27,21 @@ def to_string(exp):
 
 def repl(prompt='lis.py> '):
     "A prompt-read-eval-print loop"
+
+    special_forms = NORVIG_FORMS
+    special_forms.update(PYBUILTIN_FORMS)
+
+    global_env = add_globals(Scope(), special_forms=special_forms)
+
     while True:
         try:
-            val = eval(read(raw_input(prompt)))
+            val = eval(read(raw_input(prompt)), env=global_env)
             if val is not None: print(to_string(val))
         except Exception, e:
             exc_type, exc_value, exc_traceback = sys.exc_info()
             traceback.print_exception(exc_type, exc_value, exc_traceback)
 
 if "__main__" == __name__:
+    
+
     repl()
