+++ lispy/dialects/norvig/__init__.py
@@ -7,52 +7,13 @@ Lispy: Scheme Interpreter in Python
 
 from __future__ import division
 
-import importlib
-
-from .special_forms import NORVIG_FORMS, PYTHON_BUILTIN_FORMS
-
 
 Symbol = str
-
-SPECIAL_FORMS = NORVIG_FORMS
-SPECIAL_FORMS.update(PYTHON_BUILTIN_FORMS)
-
-class Scope(dict):
-    "A scope: a dict of {'var':val} pairs, with an outer Scope."
-    def __init__(self, parms=(), args=(), outer=None):
-        self.update(zip(parms,args))
-        self.outer = outer
-
-    def find(self, var):
-        "Find the innermost Scope where var appears."
-
-        # If ':' in var add module to environment
-        if ":" in var:
-            namespace, attribute = var.split(":")
-            module = importlib.import_module(namespace)
-
-            module_funcs = {
-                "{}:{}".format(namespace, k): attr
-                for k, attr
-                in vars(module).items()
-            }
-            self.update(module_funcs)
-        try:
-            return self if var in self else self.outer.find(var)
-        except AttributeError:
-            raise NameError("name '{}' is not defined".format(var))
-
-def add_globals(env):
-    "Add some Scheme standard procedures to an environment."
-    env.update(SPECIAL_FORMS)
-    return env
-
-global_env = add_globals(Scope())
 
 
 ################ eval
 
-def eval(x, env=global_env):
+def eval(x, env):
     "Evaluate an expression in an environment."
     if isinstance(x, Symbol):             # variable reference
         return env.find(x)[x]
